+++ src/core.ts
@@ -6,7 +6,6 @@
 import './types/modules/graphql-client.d';
 import './types/modules/mocha.parallel.d';
 import './types/modules/modifiyjs.d';
-import './types/modules/random-token.d';
 
 
 export { addRxPlugin } from './plugin';