+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Other:
   - Decreased build size by not importing `pouchdb-utils`
   - Improve build size and performance by replacing [deep-equal](https://www.npmjs.com/package/deep-equal) with [fast-deep-equal](https://github.com/epoberezkin/fast-deep-equal#readme)
+  - Remove module `random-token` and use the same random string generator everywhere.
 
 ### 10.4.0 (11 November 2021)